+++ Sources/SecondsTimer/STimerDelegate.swift
@@ -21,19 +21,19 @@
 //  SOFTWARE.
 
 public protocol STimerDelegate {
-    func clock(_ clock:STimer, didStartWithSeconds seconds:Int)
-    func clock(_ clock:STimer, didTickWithSeconds seconds:Int)
-    func clock(_ clock:STimer, didStopAtSeconds seconds:Int)
-    func clock(_ clock:STimer, didPauseAtSeconds seconds:Int)
-    func clock(_ clock:STimer, didResumeWithSeconds seconds:Int)
+    func clock(_ clock:STimer, didStartWithSeconds seconds:Double)
+    func clock(_ clock:STimer, didTickWithSeconds seconds:Double)
+    func clock(_ clock:STimer, didStopAtSeconds seconds:Double)
+    func clock(_ clock:STimer, didPauseAtSeconds seconds:Double)
+    func clock(_ clock:STimer, didResumeWithSeconds seconds:Double)
     func clockDidEnd(_ clock:STimer)
 }
 
 public extension STimerDelegate {
-    func clock(_ clock:STimer, didStartWithSeconds seconds:Int) { }
-    func clock(_ clock:STimer, didTickWithSeconds seconds:Int) { }
-    func clock(_ clock:STimer, didStopAtSeconds seconds:Int) { }
-    func clock(_ clock:STimer, didPauseAtSeconds seconds:Int) { }
-    func clock(_ clock:STimer, didResumeWithSeconds seconds:Int) { }
+    func clock(_ clock:STimer, didStartWithSeconds seconds:Double) { }
+    func clock(_ clock:STimer, didTickWithSeconds seconds:Double) { }
+    func clock(_ clock:STimer, didStopAtSeconds seconds:Double) { }
+    func clock(_ clock:STimer, didPauseAtSeconds seconds:Double) { }
+    func clock(_ clock:STimer, didResumeWithSeconds seconds:Double) { }
     func clockDidEnd(_ clock:STimer) { }
 }
